Fix: Properly delete duplicate mapsets
Former-commit-id: 5cdf62b8b8d04e8d4e55bfdab85dcfb17b142068 [formerly 488b8bf57898c9da40b646311f67b62f372489a1]
Former-commit-id: b66a1a155340ef54a26355edfe1ce85c3bb5df8f

+++ Quaver/src/Database/BeatmapCache.cs
@@ -305,7 +305,7 @@ namespace Quaver.Database
         {
             var duplicateInsertEntries = beatmaps
                 .GroupBy(x => x.Md5Checksum)
-                .Select(g => g.Last())
+                .Select(g => g.Count() > 1 ? g.First() : null)
                 .ToList();
 
             // Delete all the duplicate entries
@@ -314,6 +314,12 @@ namespace Quaver.Database
                 try
                 {
                     beatmaps.Remove(map);
+
+                    // Continue to the next map if the current one is indeed null.
+                    if (map == null)
+                        continue;
+
+                    // Delete the file if its actually a duplicate.
                     File.Delete($"{Configuration.SongDirectory}/{map.Directory}/{map.Path}");
                     Logger.Log($"Removed duplicate map: {Configuration.SongDirectory}/{map.Directory}/{map.Path}", Color.Pink);
                 }
